Replace DKDF by n-PRF in key schedule; minor fix
We propose to replace the two-stage Extract-Expand-
Extract-Expand method to derive a single key from
multiple input keys by an Expand-XOR-Expand method.

As a small fix we added the Context argument to the
Derive-Secret function and pass it on to the
HKDF-Extract-Label function. In the key schedule
graph the function was called with the Context
argument, but it wasn't present in the function
definition of Derive-Secret.

This commit was co-authored by @chrisbrzuska and
me.