Resolve remaining reference to prior_epoch
In the Sequencing section there was a reference to a prior_epoch field,
which was introduced in 2fa6ed3c6afc7d308905255b4cabe154a48d1e4b.
However, it was subsequently removed from the struct where it was added
in 69e12cd61378aae4a81199c28a16e2802666b6e3. That struct was then
removed in 27c9c28f9634a06a9b7921ea9c15e276f8c9ff6f, after which the
trail runs cold. So, I wasn't able to figure out exacly what it must be
based on what on what it originally was. Hence, I replaced it with
something that makes sense in context.

I based the new text on the following line found in the Commit section:

> Verify that the `epoch` field of the enclosing MLSPlaintext message is
> equal to the `epoch` field of the current GroupContext object