Rewrite paragraph on the `new_member` SenderType
I found the phrase " In such cases" after the sentence "Proposals with types other than Add MUST NOT be sent with this sender type" slightly confusing and awkward. To improve upon this I rewrote the entire paragraph.

This is achieved by combining the restrictions imposed due to the SenderType into one sentence AND starting the sentence describing signatures without a  sentence connector.