Use the GroupContext to derive the joiner_secret
This is a proposal to change how the `joiner_secret` is derived. Before,
it was derived simply by derivation with a label. This Pull Request
updates this step to include the GroupContext, similar to how the
`epoch_secret` is derived. Including the GroupContext will give stronger
uniqueness guarantees on the `joiner_secret` cryptographically speaking.
I think the GroupContext in the derivation of the `epoch_secret` should
be kept as it proves that joiners have learned the GroupContext through
the Welcome message as well.